# Conflicts:
#	TIGLViewer/src/TIGLViewerInternal.h
#	TIGLViewer/src/TIGLViewerWidget.cpp
#	TIGLViewer/src/TIGLViewerWindow.cpp
#	src/api/tigl.cpp
#	src/common/tiglcommonfunctions.cpp
#	src/common/tiglcommonfunctions.h
#	src/configuration/CCPACSConfiguration.cpp
#	src/configuration/CCPACSConfiguration.h
#	src/cpacs_other/CTiglUIDManager.cpp
#	src/cpacs_other/CTiglUIDManager.h
#	src/exports/CTiglExportIges.cpp
#	src/exports/CTiglExportStep.cpp
#	src/exports/CTiglExportStl.cpp
#	src/fuselage/CCPACSFuselage.cpp
#	src/fuselage/CCPACSFuselageProfile.cpp
#	src/fuselage/CCPACSFuselages.h
#	src/geometry/CFunctionToBspline.cpp
#	src/geometry/CTiglTransformation.cpp
#	src/geometry/CTiglTransformation.h
#	src/wing/CCPACSWing.cpp
#	src/wing/CCPACSWing.h
#	src/wing/CCPACSWingCell.cpp
#	src/wing/CCPACSWingCells.cpp
#	src/wing/CCPACSWingComponentSegment.cpp
#	src/wing/CCPACSWingComponentSegment.h
#	src/wing/CCPACSWingProfile.cpp
#	src/wing/CCPACSWingProfilePointList.cpp
#	src/wing/CCPACSWingProfiles.cpp
#	src/wing/CCPACSWingSection.cpp
#	src/wing/CCPACSWingSection.h
#	src/wing/CCPACSWingSectionElement.cpp
#	src/wing/CCPACSWingSegment.cpp
#	src/wing/CCPACSWingSegment.h
#	src/wing/CCPACSWings.cpp

+++ bindings/bindings_generator/matlab_generator.py
@@ -5,6 +5,7 @@ Created on Sat May  4 19:28:45 2013
 @author: Martin Sigel <martin.siggel@dlr.de>
 """
 
+from __future__ import print_function
 import os
 from datetime import date
 
@@ -22,6 +23,11 @@ class MatlabGenerator(object):
         
         name = fun_dec.method_name
         filename = name + '.m'
+
+        # if file already exist, don't overwrite - it might be handwritten
+        if os.path.isfile(filename):
+            return
+        
         fop = open(filename, 'w')
         
         string = 'function '
@@ -102,7 +108,7 @@ class MatlabGenerator(object):
                 string +='*mxGetPr(plhs[%d]) = (double)%s;\n' % (index, arg.name)
             else:
                 #pass
-                print arg.name, arg.type
+                print(arg.name, arg.type)
                 raise Exception('Conversion from "%s, %s" to mx not yet implemented' \
                     % (arg.type + '*'*arg.npointer, arg.name))    
         else:
